+++ src/sklearn_evaluation/nb/sets.py
@@ -4,8 +4,7 @@ Operations in sets
 
 
 def differences(a, b):
-    """Returns objects that are only on a, b and both
-    """
+    """Returns objects that are only on a, b and both"""
     a = set(a)
     b = set(b)