+++ src/sklearn_evaluation/report/util.py
@@ -3,6 +3,6 @@ from jinja2 import Environment, PackageLoader
 
 def jinja_env():
     env = Environment(
-        loader=PackageLoader('sklearn_evaluation', 'assets/report'),
+        loader=PackageLoader("sklearn_evaluation", "assets/report"),
     )
     return env