+++ tests/nb/conftest.py
@@ -7,11 +7,11 @@ import pytest
 
 
 def save_notebook(nb_str, path, execute=True, parameters=None):
-    nb = jupytext.reads(nb_str, fmt='py:light')
-    nb.metadata['kernelspec'] = {
-        'name': 'python3',
-        'language': 'python',
-        'display_name': 'Python 3'
+    nb = jupytext.reads(nb_str, fmt="py:light")
+    nb.metadata["kernelspec"] = {
+        "name": "python3",
+        "language": "python",
+        "display_name": "Python 3",
     }
 
     Path(path).parent.mkdir(exist_ok=True, parents=True)
@@ -39,7 +39,7 @@ print(list_)
 dict_ = {'x': 1, 'y': 2}
 dict_
 """
-    save_notebook(content, 'nb_literals.ipynb')
+    save_notebook(content, "nb_literals.ipynb")
 
 
 @pytest.fixture
@@ -57,7 +57,7 @@ print(list_)
 dict_ = {'x': 2, 'y': 3}
 dict_
 """
-    save_notebook(content, 'nb_other_literals.ipynb')
+    save_notebook(content, "nb_other_literals.ipynb")
 
 
 @pytest.fixture
@@ -71,7 +71,7 @@ plt.plot([1, 2, 3], [1, 2, 3])
 # + tags=["b"]
 42
 """
-    save_notebook(content, 'nb_plot.ipynb')
+    save_notebook(content, "nb_plot.ipynb")
 
 
 @pytest.fixture
@@ -85,7 +85,7 @@ pd.DataFrame({'a': [1,2 ,3]})
 # + tags=["b"]
 42
 """
-    save_notebook(content, 'nb_table.ipynb')
+    save_notebook(content, "nb_table.ipynb")
 
 
 @pytest.fixture
@@ -96,7 +96,7 @@ import pandas as pd
 # + tags=["a"]
 x = 1
 """
-    save_notebook(content, 'nb_no_output.ipynb')
+    save_notebook(content, "nb_no_output.ipynb")
 
 
 @pytest.fixture
@@ -107,7 +107,7 @@ import numpy as np
 # + tags=["numpy_array"]
 np.array([1, 2, 3])
 """
-    return save_notebook(content, 'nb_invalid_output.ipynb')
+    return save_notebook(content, "nb_invalid_output.ipynb")
 
 
 @pytest.fixture
@@ -121,7 +121,7 @@ x = 1
 y = [1, 2]
 z = {'a': 1, 'b': 2}
 """
-    return save_notebook(content, 'nb.ipynb', execute=False)
+    return save_notebook(content, "nb.ipynb", execute=False)
 
 
 @pytest.fixture
@@ -137,7 +137,7 @@ y = [1,
 z = {'a': 1,
      'b': 2}
 """
-    return save_notebook(content, 'nb.ipynb', execute=False)
+    return save_notebook(content, "nb.ipynb", execute=False)
 
 
 @pytest.fixture
@@ -146,7 +146,7 @@ def nb_str():
 # + tags=["str"]
 print('something')
 """
-    return save_notebook(content, 'nb.ipynb')
+    return save_notebook(content, "nb.ipynb")
 
 
 @pytest.fixture
@@ -157,7 +157,7 @@ print('a')
 print('b')
 print('c')
 """
-    return save_notebook(content, 'nb.ipynb')
+    return save_notebook(content, "nb.ipynb")
 
 
 @pytest.fixture
@@ -169,4 +169,4 @@ import sys
 print('this should not appear', file=sys.stderr)
 print('something')
 """
-    return save_notebook(content, 'nb.ipynb')
+    return save_notebook(content, "nb.ipynb")