+++ CHANGELOG.rst
@@ -1,8 +1,20 @@
 CHANGELOG
 =========
 
-0.8.5dev
+0.8.7dev
 --------
+
+0.8.6 (2023-01-04)
+------------------
+* Adds support for ``pandas.Series`` to prediction error plot (#154)
+* Fixes error to prediction plot due to hardcoded axis (#155)
+* Deprecates ``model`` argument in prediction plot
+
+0.8.5 (2022-12-29)
+------------------
+* Fixes to ``plot.elbow_curve``: no longer hardcoding figure size, random seed for model, also showing ``n_clusters`` in title
+* Fixes error when passing array to ``table.feature_importances`` (#144)
+* Adds clustering quick start
 
 0.8.4 (2022-12-24)
 ------------------